chore: cargo clippy --fix auto-fixes
机械修复，由 cargo clippy --lib --fix 自动产生，零语义影响：
- match_like_matches_macro / question_mark / redundant_pattern_matching
- needless_range_loop / manual_clamp / neg_cmp_op_on_partial_ord
- single_char_pattern / single_char_add_str / explicit_iter_loop
- unnecessary_cast / explicit_lifetimes / collapsible_else_if
- get(0) → first()，trim_end_matches 闭包 → 字符数组
- thread_local OnceCell 加 const block

不含 funding_rate/{arb_decision,strategy_loader}.rs（与 Part E 改动捆绑提交）。

+++ src/strategy/mm_hedge_strategy.rs
@@ -882,7 +882,7 @@ impl MarketMakerHedgeStrategy {
     fn send_hedge_query(&mut self) {
         // 发送对冲查询：携带 symbol、当期买/卖成交、累计净头寸和风控敞口上限。
         // 调用方已确认净敞口超过阈值；这里负责组包、记录 pending 状态并启动 watchdog。
-        let hedge_ts_ms = (get_timestamp_us() / 1000) as i64;
+        let hedge_ts_ms = get_timestamp_us() / 1000 ;
         self.last_hedge_ts_ms = Some(hedge_ts_ms);
         let risk_loader = PreTradeParamsLoader::instance();
         let open_venue = MonitorChannel::instance().open_venue();
